Revert "audio: fix compiler issue due to improper braces or indentation"
This reverts commit c4a98e012b6b5a54542a039d1fed64a15a3bfb66.

This is incorrect, as the callback function should bail out when
encountering invalid audio port index (anything other than IN or OUT)
regardless of being on valid session ID (and unlock it) or not.

Signed-off-by: Albert I <kras@raphielgang.org>

techpack/audio: q6asm: Update indentation to fix compiler warning (msm-4.9)

GCC 6+ and Clang 11+ warn whenever indentations are considered misleading
especially when they think that codes are misinterpreted as if being guarded
by conditions above them.

Signed-off-by: Albert I <kras@raphielgang.org>